refactor: move systems to model
