+++ core/src/main/kotlin/com/tdt4240/group3/Hexa_Battle.kt
@@ -4,8 +4,7 @@ import com.badlogic.gdx.graphics.g2d.BitmapFont
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
 import com.badlogic.ashley.core.Engine
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer
-import com.tdt4240.group3.controller.systems.PlayerSystem
-import com.tdt4240.group3.model.components.GameStateComponent
+import com.tdt4240.group3.model.systems.PlayerSystem
 import com.tdt4240.group3.view.screens.HowToPlayScreen
 import com.tdt4240.group3.view.systems.View
 import ktx.app.KtxGame
@@ -15,7 +14,6 @@ import com.tdt4240.group3.view.screens.MenuScreen
 import com.tdt4240.group3.view.screens.PlayScreen
 import com.tdt4240.group3.view.screens.LobbyScreen
 import com.tdt4240.group3.view.screens.OptionsScreen
-import ktx.ashley.entity
 import ktx.assets.disposeSafely
 import kotlin.math.sqrt
