+++ core/src/main/kotlin/com/tdt4240/group3/controller/PauseController.kt
@@ -1,14 +1,8 @@
 package com.tdt4240.group3.controller
 
-import com.tdt4240.group3.controller.systems.TurnSystem
-import com.tdt4240.group3.model.components.GameStateComponent
-import com.tdt4240.group3.model.components.TeamComponent
+import com.tdt4240.group3.model.systems.TurnSystem
 import com.tdt4240.group3.view.screens.PlayScreen
-import com.tdt4240.group3.view.states.PauseState
-import com.tdt4240.group3.view.states.PlayerTurnState
-import com.tdt4240.group3.view.states.EnemyTurnState
 import com.tdt4240.group3.view.states.PlaySubState
-import ktx.ashley.get
 
 class PauseController(
     private val turnSystem: TurnSystem,
