+++ core/src/main/kotlin/com/tdt4240/group3/model/systems/SelectionSystem.kt
@@ -1,4 +1,4 @@
-package com.tdt4240.group3.controller.systems
+package com.tdt4240.group3.model.systems
 
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.core.EntitySystem
