+++ core/src/main/kotlin/com/tdt4240/group3/view/screens/PlayScreen.kt
@@ -18,15 +18,13 @@ import com.tdt4240.group3.controller.PauseController
 import com.tdt4240.group3.model.components.TeamComponent
 import com.tdt4240.group3.model.entities.EntityFactory
 import com.tdt4240.group3.model.systems.CollisionSystem
-import com.tdt4240.group3.controller.systems.MovementSystem
-import com.tdt4240.group3.controller.systems.SelectionSystem
-import com.tdt4240.group3.controller.systems.TroopCreationSystem
-import com.tdt4240.group3.controller.systems.TurnSystem
+import com.tdt4240.group3.model.systems.MovementSystem
+import com.tdt4240.group3.model.systems.SelectionSystem
+import com.tdt4240.group3.model.systems.TroopCreationSystem
+import com.tdt4240.group3.model.systems.TurnSystem
 import com.tdt4240.group3.model.components.GameStateComponent
 import com.tdt4240.group3.view.states.PlaySubState
 import com.tdt4240.group3.view.states.PlayerTurnState
-import com.tdt4240.group3.view.states.EnemyTurnState
-import com.tdt4240.group3.view.states.PauseState
 import ktx.actors.onClick
 import ktx.app.KtxScreen
 import ktx.ashley.allOf
